// lib/type.ts
export  type Payload = {
  responseCode: string;
  responseMessage: string;
  responseData: Contact[];
};

export  type Contact = {
  firstName: string;
  lastName: string;
  nickName: string | null;
  phone: string;
  email: string;
  dateOfBirth: string | null;
  workPhone: string;
  company: string;
  title: string;
  website: string;
  address: string;
  contactType: string;
  photoUrl: string;
  logoUrl: string;
  note: string | null;
  active: boolean;
  cardOrientation: string;
  textColor: string;
  qrColor: string | null;
  uniqueCode: string;
  inAppBillingToken: string | null;
  font: string;
  companyId: string;
  paymentLink: string;
  instagramLink: string;
  twitterLink: string;
  linkedInLink: string;
  facebookLink: string;
  youTubeChannelLink: string;
  tikTokLink: string;
  snapChatLink: string;
  bio: string;
  covidVaccinationUrl: string;
  noOfScans: number;
  dateOfLastScan: string;
  noOfClicks: number;
  dateOfLastClick: string;
  nextPaymentDueDate: string;
  cardPaymentType: string;
  userAuthId: number;
  cardDesignTypeID: number;
  userAuth: UserAuth;
  cardDesignType: CardDesignType;
  companyPreset: CompanyPreset;
  userConnections: any; // Or a more specific type if known
  checked: boolean;
  id: number;
  dateCreated: string;
  createdBy: number;
  lastModifiedBy: string | null;
  lastModifiedDate: string;
};

export  type UserAuth = {
  firstName: string;
  lastName: string;
  emailAddress: string;
  phone: string;
  password: string;
  lastLoginDate: string;
  apiKey: string;
  emailVerified: boolean;
  active: boolean;
  userType: string;
  myReferralCode: string;
  referrerCode: string | null;
  subAccountRef: string | null;
  bankCode: string | null;
  accountNumber: string | null;
  companyId: string;
  proStatus: boolean;
  proOrderReference: string | null;
  refreshToken: string;
  refreshTokenExpiry: string;
  profileCode: string;
  profileType: string;
  lockStatus: boolean;
  unlockedDurationDate: string;
  autoLockUserProfile: boolean;
  resetOnFirstLogin: boolean;
  passwordTextForSendGrid: string | null;
  id: number;
  dateCreated: string;
  createdBy: number;
  lastModifiedBy: string | null;
  lastModifiedDate: string;
};

export  type CardDesignType = {
  id: number;
  backgroundUrlHorizontal: string;
  backgroundUrlVertical: string;
  active: boolean;
  companyId: string | null;
};

export  type CompanyPreset = {
  companyId: string;
  backgroundUrl: string;
  textColor: string;
  buttonColor: string;
  buttonTextColor: string;
  buttonBorderColor: string;
  logo: string;
  instagramLink: string;
  twitterLink: string;
  linkedInLink: string;
  facebookLink: string;
  youTubeChannelLink: string;
  tikTokLink: string;
  snapChatLink: string | null;
  company: Company;
  id: number;
  dateCreated: string;
  createdBy: number;
  lastModifiedBy: string;
  lastModifiedDate: string;
};

export  type Company = {
  companyID: string;
  companyName: string;
  companySector: string | null;
  companyContact: string;
  companyContactPhone: string;
  companyContactEmail: string;
  companyWebsite: string;
  companyAddress: string;
  businessType: string;
  dateCreated: string;
  dateOfLastPayment: string;
  active: boolean;
  companySize: string | null;
  costPerCard: number;
  costPerCardUsd: number;
  firstYearFree: boolean;
};